Reformat paper to JOSS specs and length
Copies body of the LaTeX document into a Markdown file that follows the
specifications for a JOSS example paper [1]. Cuts the tool usage section
completely to reduce length, following guidance from the JOSS submission guide.

Renames the bibliography file to match the Markdown file and adds a missing
reference for Fowler and Fields 2014.

Documentation for how to build the paper locally with pandoc is in the README.

[1] https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography