Replace default pandoc template
Adds support for the same authors and affiliations config syntax in the
headmatter and apperance as used by JOSS. Also, changes the page type to A4 to
match JOSS style. These changes are based mostly on the JOSS latex template
[1].

One additional change here that should only affect the local rendering is that
figures are forced to be placed where they occur in the text using the float
package and the "H" float specifier [2]. This change prevents the figure from
getting placed later in the PDF and orphaning the "Figures" header in the text.

[1] https://github.com/openjournals/whedon/blob/master/resources/latex.template
[2] https://www.overleaf.com/learn/latex/Positioning_of_Figures